requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
pydantic==2.8.2
pyyaml==6.0.2
jinja2==3.1.4
python-multipart==0.0.9
pytest==8.3.2
minio==7.2.10
requests==2.32.3